"Fix" command line errors by limiting to printable ASCII
Still quite a bit to do and decide on.
1) There are sometimes 'unexpected argument' errors for particular
runs --- you should try to figure out why this is happening and how to
avoid it.
2) Limiting to only printable ASCII is...very limiting. That said, so
is limiting to non-backtracking regexes so...there are some design
decisions to think on here.

We can try to just start with this and do some differential testing
even in this very limited sense between engines that already build in
non-backtracking as the default. Perhaps you'll still be able to find
some (shallow) errors.